cpufreq: schedutil: Allow driver to set default rate limit
For platforms that report a DVFS transition latency, that latency is
multiplied by 1000 to decide the default rate limit. 1ms is not an unusual
transition latency and 1s is not a sensible rate limit, so such platforms
currently lack a sane default.

The rate limit is intended to avoid thrashing DVFS and using excessive
power due to voltage jitter. The ideal limit is platform-specific, and
although it is closely related to the transition latency, that relationship
is also platform- specific, so allow the driver to directly specify the
default rate limit by setting it in the cpufreq_policy.

In order to avoid violating assumptions in existing drivers such as
acpi-cpufreq about the relationship between reported transition latency and
frequency of update requests, the default rate limit for drivers that do
not set a default rate limit is not altered.

Note that this does not guarantee the driver will not receive requests at a
higher rate, since it can be overridden by userspace.

Signed-off-by: Brendan Jackman <brendan.jackman@arm.com>