sched/fair: Add nohz idle kick for solving misfit scenarios
When a lower-capacity CPU has a large task, it sets the misfit_task flag in its
rq to signal that load balance should try to pull that task to a higher capacity
CPU. However load balance may not occur for some time if those higher capacity
CPUs are in nohz idle.

With this patch, where there is a sched_domain with SD_ASYM_CPUCAPACITY, we have
a misfit task, we kick load balance on any nohz idle CPUs with maximum capacity.

This requires adding a percpu pointer to the first SD_ASYM_CPUCAPACITY
sched_domain, and a cpumask in the root domain to store the maximum capacity
CPUs.

For simplicity and lack of testing platforms, this is is implemented with only
two CPU capacities in mind (i.e. CPUs with max capacity and everyone else). On a
system with more than two CPU capacities, the "middle" CPUs will not be kicked
to pull misfit tasks from the lowest capacity CPUs.

Signed-off-by: Brendan Jackman <brendan.jackman@arm.com>