sched/fair: Fix slow-path wakeup when not on sibling of prev_cpu
The case in question is this:

 - select_task_rq fair has been sent to the slow path by wake_wide or wake_cap.
 - prev_cpu and smp_processor_id() are not siblings, i.e. not in the same group
   on the highest level sched_domain.
 - prev_cpu's siblings are all "busy" and smp_processor_id()'s group are all
   "idle", by the measures used in find_idlest_group.

Then find_idlest_group will return NULL on each invocation, indicating that the
gorup containing 'cpu' (which is currently set to smp_process_id()) is the
idlest. This means we'll exit the 'while (sd)' loop without changing
new_cpu from its initial value, which is prev_cpu. So we'll return prev_cpu
despite having just determined that smp_processor_id() is idler.

Therefore set new_cpu to 'cpu' (which, in the case in question, equals
smp_processor_id()) when find_idlest_group returns NULL.